added LinkUpdate dto, small fix in repository

+++ scrapper/src/main/kotlin/ru/bytebosses/scrapper/domain/repository/LinkRepository.kt
@@ -6,7 +6,7 @@ import ru.bytebosses.scrapper.domain.entity.LinkEntity
 import java.time.OffsetDateTime
 import java.util.*
 
-interface LinkRepository : JpaRepository<LinkEntity, Int> {
+interface LinkRepository : JpaRepository<LinkEntity, Long> {
 
     fun findByUrl(url: String): Optional<LinkEntity>
